Fix Parsing Messages With Quotes
Not sure why I used shlex to parse the incoming message.. I think it was to parse the enclosing message within quotes. I.e. Fanoen tells you, "This is the enclosing message!!!"

But that would then blow up if there were quotes within the nested message.

With this change, we use the built in string.split(' ') instead of shlex.split and simply remove the enclosing quotes by stripping off the first/last character.